minor improvements in constraint tests
