+++ src/processor/tests/comparisons.rs
@@ -1,0 +1,115 @@
+use crate::{ ProofOptions, ProgramInputs, opcodes::f128 as opcodes, F128, FiniteField, Accumulator };
+use super::super::{ execute, verify };
+
+#[test]
+fn eq_operations() {
+    let program = [
+        opcodes::BEGIN, opcodes::EQ,     opcodes::SWAP2, opcodes::EQ,
+        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
+        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
+        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
+    ];
+    let expected_hash = <F128 as Accumulator>::digest(&program[..(program.len() - 1)]);
+
+    let options = ProofOptions::default();
+    let inputs = ProgramInputs::from_public(&[1, 2, 3, 4, 4]);
+    let num_outputs = 3;
+
+    let expected_result = vec![1, 0, 3];
+
+    let (outputs, program_hash, proof) = execute(&program, &inputs, num_outputs, &options);
+    assert_eq!(expected_result, outputs);
+    assert_eq!(program_hash, expected_hash);
+
+    let result = verify(&program_hash, inputs.get_public_inputs(), &outputs, &proof);
+    assert_eq!(Ok(true), result);
+}
+
+#[test]
+fn cmp_operation() {
+
+    let a: u128 = F128::rand();
+    let b: u128 = F128::rand();
+    let p127: u128 = F128::exp(2, 127);
+
+    // build inputs
+    let (inputs_a, inputs_b) = build_inputs_for_cmp(a, b, 128);
+
+    // build the program
+    let mut program = vec![opcodes::BEGIN, opcodes::PUSH, p127];
+    for _ in 0..128 { program.push(opcodes::CMP);  }
+    program.push(opcodes::DROP);
+    program.push(opcodes::DROP);
+    program.push(opcodes::DROP);
+    while program.len() < 256 { program.push(opcodes::NOOP); }
+
+    let expected_hash = <F128 as Accumulator>::digest(&program[..(program.len() - 1)]);
+
+    let options = ProofOptions::default();
+    let inputs = ProgramInputs::new(&[0, 0, 0, 0, 0, 0, a, b], &inputs_a, &inputs_b);
+    let num_outputs = 4;
+
+    let lt = if a < b { F128::ONE }  else { F128::ZERO };
+    let gt = if a < b { F128::ZERO } else { F128::ONE  };
+    let expected_result = vec![gt, lt, b, a];
+
+    // execute the program and make sure results are correct
+    let (outputs, program_hash, proof) = execute(&program, &inputs, num_outputs, &options);
+    assert_eq!(expected_result, outputs);
+    assert_eq!(program_hash, expected_hash);
+
+    // verify execution proof
+    let result = verify(&program_hash, inputs.get_public_inputs(), &outputs, &proof);
+    assert_eq!(Ok(true), result);
+}
+
+#[test]
+fn binacc_operation() {
+
+    let a: u128 = F128::rand();
+    let p127: u128 = F128::exp(2, 127);
+
+    // build inputs
+    let mut inputs_a = Vec::new();
+    for i in 0..128 { inputs_a.push((a >> i) & 1); }
+    inputs_a.reverse();
+
+    // build the program
+    let mut program = vec![opcodes::BEGIN];
+    for _ in 0..128 { program.push(opcodes::BINACC); }
+    program.push(opcodes::DROP);
+    while program.len() < 256 { program.push(opcodes::NOOP); }
+
+    let expected_hash = <F128 as Accumulator>::digest(&program[..(program.len() - 1)]);
+
+    let options = ProofOptions::default();
+    let inputs = ProgramInputs::new(&[p127, 0, a], &inputs_a, &[]);
+    let num_outputs = 2;
+
+    let expected_result = vec![a, a];
+
+    // execute the program and make sure results are correct
+    let (outputs, program_hash, proof) = execute(&program, &inputs, num_outputs, &options);
+    assert_eq!(expected_result, outputs);
+    assert_eq!(program_hash, expected_hash);
+
+    // verify execution proof
+    let result = verify(&program_hash, inputs.get_public_inputs(), &outputs, &proof);
+    assert_eq!(Ok(true), result);
+}
+
+// HELPER FUNCTIONS
+// ================================================================================================
+fn build_inputs_for_cmp(a: u128, b: u128, size: usize) -> (Vec<u128>, Vec<u128>) {
+
+    let mut inputs_a = Vec::new();
+    let mut inputs_b = Vec::new();
+    for i in 0..size {
+        inputs_a.push((a >> i) & 1);
+        inputs_b.push((b >> i) & 1);
+    }
+    inputs_a.reverse();
+    inputs_b.reverse();
+
+    return (inputs_a, inputs_b);
+}
