+++ src/processor/tests/mod.rs
@@ -1,5 +1,6 @@
-#[cfg(test)]
 use crate::{ ProofOptions, ProgramInputs, opcodes::f128 as opcodes, F128, FiniteField, Accumulator, Hasher };
+
+mod comparisons;
 
 #[test]
 fn execute_verify() {
@@ -245,67 +246,6 @@ fn read_operations() {
     assert_eq!(Ok(true), result);
 }
 
-#[test]
-fn comparison_operations() {
-    let program = [
-        opcodes::BEGIN, opcodes::EQ,     opcodes::SWAP2, opcodes::EQ,
-        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
-        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
-        opcodes::NOOP,  opcodes::NOOP,   opcodes::NOOP,  opcodes::NOOP,
-    ];
-    let expected_hash = <F128 as Accumulator>::digest(&program[..(program.len() - 1)]);
-
-    let options = ProofOptions::default();
-    let inputs = ProgramInputs::from_public(&[1, 2, 3, 4, 4]);
-    let num_outputs = 3;
-
-    let expected_result = vec![1, 0, 3];
-
-    let (outputs, program_hash, proof) = super::execute(&program, &inputs, num_outputs, &options);
-    assert_eq!(expected_result, outputs);
-    assert_eq!(program_hash, expected_hash);
-
-    let result = super::verify(&program_hash, inputs.get_public_inputs(), &outputs, &proof);
-    assert_eq!(Ok(true), result);
-}
-
-#[test]
-fn cmp_operation() {
-    // TODO: improve
-    let a: u128 = F128::rand();
-    let b: u128 = F128::rand();
-    let p127: u128 = F128::exp(2, 127);
-
-    let mut inputs_a = Vec::new();
-    let mut inputs_b = Vec::new();
-    for i in 0..128 {
-        inputs_a.push((a >> i) & 1);
-        inputs_b.push((b >> i) & 1);
-    }
-    inputs_a.reverse();
-    inputs_b.reverse();
-
-    let mut program = vec![opcodes::BEGIN, opcodes::PUSH, p127];
-    for _ in 0..128 { program.push(opcodes::CMP);  }
-    while program.len() < 256 { program.push(opcodes::NOOP); }
-
-    let expected_hash = <F128 as Accumulator>::digest(&program[..(program.len() - 1)]);
-
-    let options = ProofOptions::default();
-    let inputs = ProgramInputs::new(&[0, 0, 0, 0, 0, 0, a, b], &inputs_a, &inputs_b);
-    let num_outputs = 8;
-
-    let lt = if a < b { F128::ONE }  else { F128::ZERO };
-    let gt = if a < b { F128::ZERO } else { F128::ONE  };
-    let expected_result = vec![gt, lt, b, a];
-
-    let (outputs, program_hash, proof) = super::execute(&program, &inputs, num_outputs, &options);
-    assert_eq!(expected_result, outputs[3..7].to_vec());
-    assert_eq!(program_hash, expected_hash);
-
-    let result = super::verify(&program_hash, inputs.get_public_inputs(), &outputs, &proof);
-    assert_eq!(Ok(true), result);
-}
 
 #[test]
 fn assert_operation() {
